Strip "local" bits from system paths.
Starting with Python 3.10 Debian changed how it reports paths via sysconfig.
They now have a 'local' path segment in them, presumably to match that
when installed system-wide things to go "/usr/local". When you install in
a custom dir, though, the "local" part is not actually used. So don't use that
then.